Fix parsing of pin numbers
Not all pin names found by the script are as nice as it used to think.
Some have extra characters after the number, causing it to fail when
trying to convert the pin-number-plus-extra-characters to an int.  Now
we use a regex to get the pin number, which is much more reliable.