Fix parsing STM32 RAM and Flash sizes
They previously only looked at the first Ram and Flash elements in the
XML file, which caused inaccurate documentation for parts that share a
single XML file (e.g. STM32F072C(8-B)Ux.xml).  This is fixed now, so a
lot of parts will have documentation fixes when this branch is merged.