Use f-strings to build parts
Lots of string concatenation with + is generally considered bad, since
it allocates and destroys way more strings than should be allocated and
destroyed.  Now we format lines as f-strings, append them to lists, and
join the lists when we're done.  The result is the same, but it's a much
nicer way of getting there.