Use the KiCadSymbolGenerator package for STM32 lib
We have a nice symbol generator package, why not use it?  This is a
major rewrite of the script, with a somewhat different symbol shape (a
bit wider and shorter).  All the same information is there, but the
symbols are generated more nicely, with Python objects for drawing
elements and such.