Added a primitive check for stencil pads
This commit improves checking of rule F6.3 for the case of thermal pads
with separate stencil opening pads.  If a pad is found to have a missing
paste layer, we now check if there are any stencil pads entirely inside
the area of the thermal pad.

The implementation here only works for non-rotated rectangular thermal
pads and paste pads.  This isn't ideal, but should cover the vast
majority of thermal pads in the library.  Any non-supported pads are
ignored by this new code path, so it should only remove false errors.