Correctly read footprint's clearance setting
There was a problem before where a KicadMod's clearance attribute was
set to that of a pad (or the clearance option of a custom shaped pad) if
the footprint itself didn't have a clearance value set.  This ultimately
was caused by improper recursion depth limiting in KicadMod._getArray.

This commit fixes the problem, making the attributes read from only the
top level of the kicad_mod's module object.