Position top/bottom pins correctly
Before, the top and bottom pin groups were centered on the body
rectangle.  This caused pin names to overlap in some parts, e.g.
STM32F437IGTx.  This commit positions the top and bottom pin groups in
the center of the space between the left- and right-side pin names,
fixing the problem.