Give an info message for pin offset > 20 mil
KLC says that the preferred pin name position offset is 20 mil, and that
larger values up to 50 mil are allowed for unusual symbol geometry.
This commit causes an info message to be printed if the offset is > 20
mil, because I keep failing to catch that for rectangular symbols where
it should be 20 mil.