S3.6: Allow offset to be 0
An offset of 0 in the library file is special: it means the pin name is
next to the pin rather than inside the symbol.  In some cases, that's
the only way to draw the pin name, so there's nothing wrong if we're in
that case.  This commit makes the script silent when the offset is set
to 0.