Only check S3.6 if pin names are drawn
Rule S3.6 should only apply if pin names are drawn; trying to check it
if they aren't drawn doesn't make sense.  This would be a KLC change
that hasn't been discussed yet, so I'd like to see more than one other
maintainer agree before merging this!

This commit causes the S3.6 check to be skipped if pin names are not
drawn.