Issue a warning for missing footprint field
As pointed out in #234, there are plenty of cases where a symbol having
one footprint filter and no default footprint is valid, and therefore
this shouldn't be treated as an error by the script.

This commit changes that case to a warning, splits up the message into a
warning and a warningExtra, and shortens the message overall by not
explicitly saying how to fix the problem if it in fact is a problem.