Moved extensions in F9.3 check
The variable `extensions` holds all the allowed extensions for 3D
models.  This is only used for a membership test, so changing its type
from a list to a set makes sense (testing membership in sets is faster).
Also, it makes sense to move the definition of `extensions` to where
it's actually used, rather than burying it in an unrelated part of the
check method.