Remove redundant call to str.startswith
When checking if a 3D model name matches part of the footprint name or
vice versa, we first checked if the fp_name.startswith(model_file).
This is redundant, since we do the more general check with the in
operator immediately afterwards.  Also, quick testing with the timeit
module shows that str.startswith is slower than the in operator by
roughly a factor of four.  Accordingly, this commit removes the call to
str.startswith.