Improve unidiomatic tests
Many tests in F9.3 didn't follow Python idioms: strange use of "not"
(not x in y -> x not in y; not x == y -> x != y) and odd breaking of
chained conditions (using backslashes for continuation rather than
parentheses, "or" at the end of the line rather than the beginning of
the next).  This fixes these to be more Pythonic.