Detect and fix footprint suffixes in 3D models
Certain suffixes, notably _ThermalVias, _Pad*x*mm, _HandSolder, and
_CircularHoles, don't affect the 3D model.  KLC rule F9.3 says that such
suffixes shouldn't be in the 3D model name, as this would cause
unnecessary duplication of 3D models.

Evan pointed out in #246 that the script could look for these and not
warn if they're missing from the 3D model name.  I decided to take it a
step further, and make it warn if they are present.  This commit does
that, and makes the script able to automatically remove these fields
from the 3D model name with --fixmore.