Add rouding to mmToMicrons() function
For when using getStartPoint() on arcs combined with the mmToMicrons()
function the accuracy was not good leading to errors of the F5.3 check.
See https://github.com/KiCad/kicad-footprints/pull/2267

With rounding applied to the value before converting it to int that issue
is gone. This change won't break other rules since the function mmToMicrons()
is only used for rule F5.3.

Fixes #329